Add Yocto Krogoth support
As of March 31, 2023, all codeaurora.org repositories have been migrated
to other platforms and the project has been shut down.
For more information, please visit: https://bye.codeaurora.org/

Variscite's Krogoth releases were based on a NXP's manifest, but NXP didn't
update this manifest and some of the meta layers are still pointing to CAF,
which causes errors during repo init.

Create manifests for Krogoth releases, based on NXP's manifest and fix the
reference to CAF repositories. Also, add meta-variscite-imx to the manifest
to make it easier to build the release, instead of manually cloning the
meta layer and copying the script to initialize the build directory.

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>